Don't cache MongoLockProvider in the integration test
The properties tested in the abstract test suite should still hold, even when the lock provider is created on demand (it's synonymous to multiple clients trying to acquire the lock). Unfortunately, when only a single instance is used in the test, it's easy to make an uncaught mistake when writing a custom implementation. In my case I've made an error of using regular `set` updates in my `insertRecord` implementation (instead of correct `setOnInsert`) and the tests still passed just fine.